+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 2021-10-08
+### Changed
+- Use Fabric API's TriState for the Fabric swimming event
+
+### Fixed
+- Depend on Minecraft 1.17+ instead of 1.17.1+
+
 ## [3.0.2] - 2021-10-04
 ### Fixed
 - Pulled in fix from 1.16 version (2.0.2)
@@ -48,6 +55,7 @@ Initial release
 ### Added
 - Player swim event
 
+[4.0.0]: https://github.com/florensie/artifacts-fabric/compare/v3.0.2..v4.0.0
 [3.0.2]: https://github.com/florensie/artifacts-fabric/compare/v3.0.1..v3.0.2
 [2.0.2]: https://github.com/florensie/artifacts-fabric/compare/v2.0.1..v2.0.2
 [3.0.1]: https://github.com/florensie/artifacts-fabric/compare/v3.0.0..v3.0.1